refactor: added check for target existence

+++ kura/distrib/src/main/resources/nvidia-jetson-nano-nn/kura_install.sh
@@ -20,7 +20,7 @@ setup_libudev() {
         ln -sf /lib/libudev.so.1 /lib/libudev.so.0
     fi
 
-    if [ ! -f /usr/lib/aarch64-linux-gnu/libudev.so.0 ]; then
+    if [ ! -f /usr/lib/aarch64-linux-gnu/libudev.so.0 ] && [ -f /usr/lib/aarch64-linux-gnu/libudev.so.1 ]; then
        ln -sf /usr/lib/aarch64-linux-gnu/libudev.so.1 /usr/lib/aarch64-linux-gnu/libudev.so.0
     fi
 }
